Move api-specific utils
